refactor outputs
