+++ prowler/lib/outputs/outputs.py
@@ -1,46 +1,35 @@
 import json
-import os
 import sys
 from csv import DictWriter
 from io import TextIOWrapper
 from typing import Any
 
 from colorama import Fore, Style
-from tabulate import tabulate
 
 from prowler.config.config import (
     csv_file_suffix,
     html_file_suffix,
-    html_logo_img,
-    html_logo_url,
     json_asff_file_suffix,
     json_file_suffix,
     orange_color,
-    prowler_version,
     timestamp,
-    timestamp_utc,
 )
 from prowler.lib.logger import logger
+from prowler.lib.outputs.csv import generate_csv_fields
+from prowler.lib.outputs.html import add_html_header, fill_html
+from prowler.lib.outputs.json import fill_json_asff
 from prowler.lib.outputs.models import (
     Aws_Check_Output_CSV,
     Azure_Check_Output_CSV,
     Check_Output_CSV_CIS,
     Check_Output_CSV_ENS_RD2022,
     Check_Output_JSON_ASFF,
-    Compliance,
-    ProductFields,
-    Resource,
-    Severity,
-    generate_csv_fields,
-    generate_provider_output_csv,
-    generate_provider_output_json,
 )
-from prowler.lib.utils.utils import file_exists, hash_sha512, open_file
+from prowler.lib.utils.utils import file_exists, open_file
 from prowler.providers.aws.lib.allowlist.allowlist import is_allowlisted
 from prowler.providers.aws.lib.audit_info.models import AWS_Audit_Info
 from prowler.providers.aws.lib.security_hub.security_hub import send_to_security_hub
 from prowler.providers.azure.lib.audit_info.models import Azure_Audit_Info
-from prowler.providers.common.outputs import Provider_Output_Options
 
 
 def stdout_report(finding, color, verbose, is_quiet):
@@ -429,67 +418,6 @@ def set_report_color(status: str) -> str:
     return color
 
 
-def fill_json_asff(finding_output, audit_info, finding):
-    # Check if there are no resources in the finding
-    if finding.resource_id == "":
-        finding.resource_id = "NONE_PROVIDED"
-    finding_output.Id = f"prowler-{finding.check_metadata.CheckID}-{audit_info.audited_account}-{finding.region}-{hash_sha512(finding.resource_id)}"
-    finding_output.ProductArn = f"arn:{audit_info.audited_partition}:securityhub:{finding.region}::product/prowler/prowler"
-    finding_output.ProductFields = ProductFields(
-        ProviderVersion=prowler_version, ProwlerResourceName=finding.resource_id
-    )
-    finding_output.GeneratorId = "prowler-" + finding.check_metadata.CheckID
-    finding_output.AwsAccountId = audit_info.audited_account
-    finding_output.Types = finding.check_metadata.CheckType
-    finding_output.FirstObservedAt = (
-        finding_output.UpdatedAt
-    ) = finding_output.CreatedAt = timestamp_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
-    finding_output.Severity = Severity(Label=finding.check_metadata.Severity.upper())
-    finding_output.Title = finding.check_metadata.CheckTitle
-    finding_output.Description = finding.check_metadata.Description
-    finding_output.Resources = [
-        Resource(
-            Id=finding.resource_id,
-            Type=finding.check_metadata.ResourceType,
-            Partition=audit_info.audited_partition,
-            Region=finding.region,
-        )
-    ]
-    # Add ED to PASS or FAIL (PASSED/FAILED)
-    finding_output.Compliance = Compliance(
-        Status=finding.status + "ED",
-        RelatedRequirements=finding.check_metadata.CheckType,
-    )
-    finding_output.Remediation = {
-        "Recommendation": finding.check_metadata.Remediation.Recommendation
-    }
-
-    return finding_output
-
-
-def close_json(output_filename, output_directory, mode):
-    try:
-        suffix = json_file_suffix
-        if mode == "json-asff":
-            suffix = json_asff_file_suffix
-        filename = f"{output_directory}/{output_filename}{suffix}"
-        file_descriptor = open_file(
-            filename,
-            "a",
-        )
-        # Replace last comma for square bracket if not empty
-        if file_descriptor.tell() > 0:
-            file_descriptor.seek(file_descriptor.tell() - 1, os.SEEK_SET)
-            file_descriptor.truncate()
-            file_descriptor.write("]")
-        file_descriptor.close()
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}] -- {error}"
-        )
-        sys.exit()
-
-
 def send_to_s3_bucket(
     output_filename, output_directory, output_mode, output_bucket, audit_session
 ):
@@ -513,579 +441,6 @@ def send_to_s3_bucket(
             f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}] -- {error}"
         )
         sys.exit()
-
-
-def display_summary_table(
-    findings: list,
-    audit_info,
-    output_options: Provider_Output_Options,
-    provider: str,
-):
-    output_directory = output_options.output_directory
-    output_filename = output_options.output_filename
-    try:
-        if provider == "aws":
-            entity_type = "Account"
-            audited_entities = audit_info.audited_account
-        elif provider == "azure":
-            if audit_info.identity.domain:
-                entity_type = "Tenant Domain"
-                audited_entities = audit_info.identity.domain
-            else:
-                entity_type = "Tenant ID/s"
-                audited_entities = " ".join(audit_info.identity.tenant_ids)
-
-        if findings:
-            current = {
-                "Service": "",
-                "Provider": "",
-                "Total": 0,
-                "Critical": 0,
-                "High": 0,
-                "Medium": 0,
-                "Low": 0,
-            }
-            findings_table = {
-                "Provider": [],
-                "Service": [],
-                "Status": [],
-                "Critical": [],
-                "High": [],
-                "Medium": [],
-                "Low": [],
-            }
-            pass_count = fail_count = 0
-            for finding in findings:
-                # If new service and not first, add previous row
-                if (
-                    current["Service"] != finding.check_metadata.ServiceName
-                    and current["Service"]
-                ):
-
-                    add_service_to_table(findings_table, current)
-
-                    current["Total"] = current["Critical"] = current["High"] = current[
-                        "Medium"
-                    ] = current["Low"] = 0
-
-                current["Service"] = finding.check_metadata.ServiceName
-                current["Provider"] = finding.check_metadata.Provider
-
-                current["Total"] += 1
-                if finding.status == "PASS":
-                    pass_count += 1
-                elif finding.status == "FAIL":
-                    fail_count += 1
-                    if finding.check_metadata.Severity == "critical":
-                        current["Critical"] += 1
-                    elif finding.check_metadata.Severity == "high":
-                        current["High"] += 1
-                    elif finding.check_metadata.Severity == "medium":
-                        current["Medium"] += 1
-                    elif finding.check_metadata.Severity == "low":
-                        current["Low"] += 1
-
-            # Add final service
-
-            add_service_to_table(findings_table, current)
-
-            print("\nOverview Results:")
-            overview_table = [
-                [
-                    f"{Fore.RED}{round(fail_count/len(findings)*100, 2)}% ({fail_count}) Failed{Style.RESET_ALL}",
-                    f"{Fore.GREEN}{round(pass_count/len(findings)*100, 2)}% ({pass_count}) Passed{Style.RESET_ALL}",
-                ]
-            ]
-            print(tabulate(overview_table, tablefmt="rounded_grid"))
-
-            print(
-                f"\n{entity_type} {Fore.YELLOW}{audited_entities}{Style.RESET_ALL} Scan Results (severity columns are for fails only):"
-            )
-            if provider == "azure":
-                print(
-                    f"\nSubscriptions scanned: {Fore.YELLOW}{' '.join(audit_info.identity.subscriptions.keys())}{Style.RESET_ALL}"
-                )
-            print(tabulate(findings_table, headers="keys", tablefmt="rounded_grid"))
-            print(
-                f"{Style.BRIGHT}* You only see here those services that contains resources.{Style.RESET_ALL}"
-            )
-            print("\nDetailed results are in:")
-            if "html" in output_options.output_modes:
-                print(f" - HTML: {output_directory}/{output_filename}.html")
-            if "json-asff" in output_options.output_modes:
-                print(f" - JSON-ASFF: {output_directory}/{output_filename}.asff.json")
-            if "csv" in output_options.output_modes:
-                print(f" - CSV: {output_directory}/{output_filename}.csv")
-            if "json" in output_options.output_modes:
-                print(f" - JSON: {output_directory}/{output_filename}.json")
-
-        else:
-            print(
-                f"\n {Style.BRIGHT}There are no findings in {entity_type} {Fore.YELLOW}{audited_entities}{Style.RESET_ALL}\n"
-            )
-
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}:{error.__traceback__.tb_lineno} -- {error}"
-        )
-        sys.exit()
-
-
-def add_service_to_table(findings_table, current):
-    if (
-        current["Critical"] > 0
-        or current["High"] > 0
-        or current["Medium"] > 0
-        or current["Low"] > 0
-    ):
-        total_fails = (
-            current["Critical"] + current["High"] + current["Medium"] + current["Low"]
-        )
-        current["Status"] = f"{Fore.RED}FAIL ({total_fails}){Style.RESET_ALL}"
-    else:
-        current["Status"] = f"{Fore.GREEN}PASS ({current['Total']}){Style.RESET_ALL}"
-    findings_table["Provider"].append(current["Provider"])
-    findings_table["Service"].append(current["Service"])
-    findings_table["Status"].append(current["Status"])
-    findings_table["Critical"].append(
-        f"{Fore.LIGHTRED_EX}{current['Critical']}{Style.RESET_ALL}"
-    )
-    findings_table["High"].append(f"{Fore.RED}{current['High']}{Style.RESET_ALL}")
-    findings_table["Medium"].append(
-        f"{Fore.YELLOW}{current['Medium']}{Style.RESET_ALL}"
-    )
-    findings_table["Low"].append(f"{Fore.BLUE}{current['Low']}{Style.RESET_ALL}")
-
-
-def display_compliance_table(
-    findings: list,
-    bulk_checks_metadata: dict,
-    compliance_framework: str,
-    output_filename: str,
-    output_directory: str,
-):
-    try:
-        if "ens_rd2022_aws" in compliance_framework:
-            marcos = {}
-            ens_compliance_table = {
-                "Proveedor": [],
-                "Marco/Categoria": [],
-                "Estado": [],
-                "PYTEC": [],
-                "Alto": [],
-                "Medio": [],
-                "Bajo": [],
-            }
-            pass_count = fail_count = 0
-            for finding in findings:
-                check = bulk_checks_metadata[finding.check_metadata.CheckID]
-                check_compliances = check.Compliance
-                for compliance in check_compliances:
-                    if (
-                        compliance.Framework == "ENS"
-                        and compliance.Provider == "AWS"
-                        and compliance.Version == "RD2022"
-                    ):
-                        compliance_version = compliance.Version
-                        compliance_fm = compliance.Framework
-                        compliance_provider = compliance.Provider
-                        for requirement in compliance.Requirements:
-                            for attribute in requirement.Attributes:
-                                marco_categoria = (
-                                    f"{attribute['Marco']}/{attribute['Categoria']}"
-                                )
-                                # Check if Marco/Categoria exists
-                                if marco_categoria not in marcos:
-                                    marcos[marco_categoria] = {
-                                        "Estado": f"{Fore.GREEN}CUMPLE{Style.RESET_ALL}",
-                                        "Pytec": 0,
-                                        "Alto": 0,
-                                        "Medio": 0,
-                                        "Bajo": 0,
-                                    }
-                                if finding.status == "FAIL":
-                                    fail_count += 1
-                                    marcos[marco_categoria][
-                                        "Estado"
-                                    ] = f"{Fore.RED}NO CUMPLE{Style.RESET_ALL}"
-                                elif finding.status == "PASS":
-                                    pass_count += 1
-                                if attribute["Nivel"] == "pytec":
-                                    marcos[marco_categoria]["Pytec"] += 1
-                                elif attribute["Nivel"] == "alto":
-                                    marcos[marco_categoria]["Alto"] += 1
-                                elif attribute["Nivel"] == "medio":
-                                    marcos[marco_categoria]["Medio"] += 1
-                                elif attribute["Nivel"] == "bajo":
-                                    marcos[marco_categoria]["Bajo"] += 1
-
-            # Add results to table
-            for marco in marcos:
-                ens_compliance_table["Proveedor"].append("aws")
-                ens_compliance_table["Marco/Categoria"].append(marco)
-                ens_compliance_table["Estado"].append(marcos[marco]["Estado"])
-                ens_compliance_table["PYTEC"].append(
-                    f"{Fore.LIGHTRED_EX}{marcos[marco]['Pytec']}{Style.RESET_ALL}"
-                )
-                ens_compliance_table["Alto"].append(
-                    f"{Fore.RED}{marcos[marco]['Alto']}{Style.RESET_ALL}"
-                )
-                ens_compliance_table["Medio"].append(
-                    f"{Fore.YELLOW}{marcos[marco]['Medio']}{Style.RESET_ALL}"
-                )
-                ens_compliance_table["Bajo"].append(
-                    f"{Fore.BLUE}{marcos[marco]['Bajo']}{Style.RESET_ALL}"
-                )
-            if fail_count + pass_count < 0:
-                print(
-                    f"\n {Style.BRIGHT}There are no resources for {Fore.YELLOW}{compliance_fm} {compliance_version} - {compliance_provider}{Style.RESET_ALL}.\n"
-                )
-            else:
-                print(
-                    f"\nEstado de Cumplimiento de {Fore.YELLOW}{compliance_fm} {compliance_version} - {compliance_provider}{Style.RESET_ALL}:"
-                )
-                overview_table = [
-                    [
-                        f"{Fore.RED}{round(fail_count/(fail_count+pass_count)*100, 2)}% ({fail_count}) NO CUMPLE{Style.RESET_ALL}",
-                        f"{Fore.GREEN}{round(pass_count/(fail_count+pass_count)*100, 2)}% ({pass_count}) CUMPLE{Style.RESET_ALL}",
-                    ]
-                ]
-                print(tabulate(overview_table, tablefmt="rounded_grid"))
-                print(
-                    f"\nResultados de {Fore.YELLOW}{compliance_fm} {compliance_version} - {compliance_provider}{Style.RESET_ALL}:"
-                )
-                print(
-                    tabulate(
-                        ens_compliance_table, headers="keys", tablefmt="rounded_grid"
-                    )
-                )
-                print(
-                    f"{Style.BRIGHT}* Solo aparece el Marco/Categoria que contiene resultados.{Style.RESET_ALL}"
-                )
-                print("\nResultados detallados en:")
-                print(
-                    f" - CSV: {output_directory}/{output_filename}_{compliance_framework[0]}.csv\n"
-                )
-        if "cis" in str(compliance_framework):
-            sections = {}
-            cis_compliance_table = {
-                "Provider": [],
-                "Section": [],
-                "Level 1": [],
-                "Level 2": [],
-            }
-            pass_count = fail_count = 0
-            for finding in findings:
-                check = bulk_checks_metadata[finding.check_metadata.CheckID]
-                check_compliances = check.Compliance
-                for compliance in check_compliances:
-                    if compliance.Framework == "CIS-AWS" and compliance.Version in str(
-                        compliance_framework
-                    ):
-                        compliance_version = compliance.Version
-                        compliance_fm = compliance.Framework
-                        for requirement in compliance.Requirements:
-                            for attribute in requirement.Attributes:
-                                section = attribute["Section"]
-                                # Check if Section exists
-                                if section not in sections:
-                                    sections[section] = {
-                                        "Status": f"{Fore.GREEN}PASS{Style.RESET_ALL}",
-                                        "Level 1": {"FAIL": 0, "PASS": 0},
-                                        "Level 2": {"FAIL": 0, "PASS": 0},
-                                    }
-                                if finding.status == "FAIL":
-                                    fail_count += 1
-                                elif finding.status == "PASS":
-                                    pass_count += 1
-                                if attribute["Profile"] == "Level 1":
-                                    if finding.status == "FAIL":
-                                        sections[section]["Level 1"]["FAIL"] += 1
-                                    else:
-                                        sections[section]["Level 1"]["PASS"] += 1
-                                elif attribute["Profile"] == "Level 2":
-                                    if finding.status == "FAIL":
-                                        sections[section]["Level 2"]["FAIL"] += 1
-                                    else:
-                                        sections[section]["Level 2"]["PASS"] += 1
-
-            # Add results to table
-            sections = dict(sorted(sections.items()))
-            for section in sections:
-                cis_compliance_table["Provider"].append("aws")
-                cis_compliance_table["Section"].append(section)
-                if sections[section]["Level 1"]["FAIL"] > 0:
-                    cis_compliance_table["Level 1"].append(
-                        f"{Fore.RED}FAIL({sections[section]['Level 1']['FAIL']}){Style.RESET_ALL}"
-                    )
-                else:
-                    cis_compliance_table["Level 1"].append(
-                        f"{Fore.GREEN}PASS({sections[section]['Level 1']['PASS']}){Style.RESET_ALL}"
-                    )
-                if sections[section]["Level 2"]["FAIL"] > 0:
-                    cis_compliance_table["Level 2"].append(
-                        f"{Fore.RED}FAIL({sections[section]['Level 2']['FAIL']}){Style.RESET_ALL}"
-                    )
-                else:
-                    cis_compliance_table["Level 2"].append(
-                        f"{Fore.GREEN}PASS({sections[section]['Level 2']['PASS']}){Style.RESET_ALL}"
-                    )
-            if fail_count + pass_count < 0:
-                print(
-                    f"\n {Style.BRIGHT}There are no resources for {Fore.YELLOW}{compliance_fm}-{compliance_version}{Style.RESET_ALL}.\n"
-                )
-            else:
-                print(
-                    f"\nCompliance Status of {Fore.YELLOW}{compliance_fm}-{compliance_version}{Style.RESET_ALL} Framework:"
-                )
-                overview_table = [
-                    [
-                        f"{Fore.RED}{round(fail_count/(fail_count+pass_count)*100, 2)}% ({fail_count}) FAIL{Style.RESET_ALL}",
-                        f"{Fore.GREEN}{round(pass_count/(fail_count+pass_count)*100, 2)}% ({pass_count}) PASS{Style.RESET_ALL}",
-                    ]
-                ]
-                print(tabulate(overview_table, tablefmt="rounded_grid"))
-                print(
-                    f"\nFramework {Fore.YELLOW}{compliance_fm}-{compliance_version}{Style.RESET_ALL} Results:"
-                )
-                print(
-                    tabulate(
-                        cis_compliance_table, headers="keys", tablefmt="rounded_grid"
-                    )
-                )
-                print(
-                    f"{Style.BRIGHT}* Only sections containing results appear.{Style.RESET_ALL}"
-                )
-                print("\nDetailed Results in:")
-                print(
-                    f" - CSV: {output_directory}/{output_filename}_{compliance_framework[0]}.csv\n"
-                )
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}:{error.__traceback__.tb_lineno} -- {error}"
-        )
-        sys.exit()
-
-
-def add_html_header(file_descriptor, audit_info):
-    try:
-        if not audit_info.profile:
-            audit_info.profile = "ENV"
-        if isinstance(audit_info.audited_regions, list):
-            audited_regions = " ".join(audit_info.audited_regions)
-        elif not audit_info.audited_regions:
-            audited_regions = "All Regions"
-        else:
-            audited_regions = audit_info.audited_regions
-        file_descriptor.write(
-            """
-        <!DOCTYPE html>
-    <html lang="en">
-    <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <!-- Required meta tags -->
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <style>
-        .read-more {
-            color: #00f;
-        }
-
-        .bg-success-custom {
-            background-color: #98dea7 !important;
-        }
-
-        .bg-danger {
-            background-color: #f28484 !important;
-        }
-    </style>
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css"
-        integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-    <!-- https://datatables.net/download/index with jQuery, DataTables, Buttons, SearchPanes, and Select //-->
-    <link rel="stylesheet" type="text/css"
-        href="https://cdn.datatables.net/v/dt/jqc-1.12.4/dt-1.10.25/b-1.7.1/sp-1.4.0/sl-1.3.3/datatables.min.css" />
-    <link rel="stylesheet" href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css"
-        integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" crossorigin="anonymous" />
-    <style>
-        .show-read-more .more-text {
-            display: none;
-        }
-
-        .dataTable {
-            font-size: 14px;
-        }
-
-        .container-fluid {
-            font-size: 14px;
-        }
-
-        .float-left {
-            float: left !important;
-            max-width: 100%;
-        }
-    </style>
-    <title>Prowler - The Handy Cloud Security Tool</title>
-    </head>
-    <body>
-    <div class="container-fluid">
-        <div class="row mt-3">
-        <div class="col-md-4">
-            <a href="""
-            + html_logo_url
-            + """><img class="float-left card-img-left mt-4 mr-4 ml-4"
-                        src="""
-            + html_logo_img
-            + """
-                        alt="prowler-logo"></a>
-            <div class="card">
-            <div class="card-header">
-                Report Information
-            </div>
-            <ul class="list-group list-group-flush">
-                <li class="list-group-item">
-                <div class="row">
-                    <div class="col-md-auto">
-                    <b>Version:</b> """
-            + prowler_version
-            + """
-                    </div>
-                </div>
-                </li>
-                <li class="list-group-item">
-                <b>Parameters used:</b> """
-            + " ".join(sys.argv[1:])
-            + """
-                </li>
-                <li class="list-group-item">
-                <b>Date:</b> """
-            + timestamp.isoformat()
-            + """
-                </li>
-            </ul>
-            </div>
-        </div>
-        <div class="col-md-2">
-                <div class="card">
-                    <div class="card-header">
-                        AWS Assessment Summary
-                    </div>
-                    <ul class="list-group list-group-flush">
-                        <li class="list-group-item">
-                            <b>AWS Account:</b> """
-            + audit_info.audited_account
-            + """
-                        </li>
-                        <li class="list-group-item">
-                            <b>AWS-CLI Profile:</b> """
-            + audit_info.profile
-            + """
-                        </li>
-                        <li class="list-group-item">
-                            <b>Audited Regions:</b> """
-            + audited_regions
-            + """
-                        </li>
-                    </ul>
-                </div>
-            </div>
-            <div class="col-md-4">
-                <div class="card">
-                    <div class="card-header">
-                        AWS Credentials
-                    </div>
-                    <ul class="list-group list-group-flush">
-                        <li class="list-group-item">
-                            <b>User Id:</b> """
-            + audit_info.audited_user_id
-            + """
-                        </li>
-                        <li class="list-group-item">
-                            <b>Caller Identity ARN:</b>
-                            """
-            + audit_info.audited_identity_arn
-            + """
-                        </li>
-                    </ul>
-                </div>
-            </div>
-            <div class="col-md-2">
-            <div class="card">
-                <div class="card-header">
-                    Assessment Overview
-                </div>
-                <ul class="list-group list-group-flush">
-                    <li class="list-group-item">
-                        <b>Total Findings:</b> TOTAL_FINDINGS
-                    </li>
-                    <li class="list-group-item">
-                        <b>Passed:</b> TOTAL_PASS
-                    </li>
-                    <li class="list-group-item">
-                        <b>Failed:</b> TOTAL_FAIL
-                    </li>
-                    <li class="list-group-item">
-                        <b>Total Resources:</b> TOTAL_RESOURCES
-                    </li>
-                </ul>
-            </div>
-        </div>
-        </div>
-        </div>
-        <div class="row mt-3">
-        <div class="col-md-12">
-            <table class="table compact stripe row-border ordering" id="findingsTable" data-order='[[ 5, "asc" ]]' data-page-length='100'>
-            <thead class="thead-light">
-                <tr>
-                    <th scope="col">Status</th>
-                    <th scope="col">Severity</th>
-                    <th style="width:5%" scope="col">Service Name</th>
-                    <th scope="col">Region</th>
-                    <th style="width:20%" scope="col">Check Title</th>
-                    <th scope="col">Resource ID</th>
-                    <th style="width:15%" scope="col">Check Description</th>
-                    <th scope="col">Check ID</th>
-                    <th scope="col">Status Extended</th>
-                    <th scope="col">Risk</th>
-                    <th scope="col">Recomendation</th>
-                    <th style="width:5%" scope="col">Recomendation URL</th>
-                </tr>
-            </thead>
-            <tbody>
-    """
-        )
-    except Exception as error:
-        logger.error(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-        )
-
-
-def fill_html(file_descriptor, finding):
-    row_class = "p-3 mb-2 bg-success-custom"
-    if finding.status == "INFO":
-        row_class = "table-info"
-    elif finding.status == "FAIL":
-        row_class = "table-danger"
-    elif finding.status == "WARNING":
-        row_class = "table-warning"
-    file_descriptor.write(
-        f"""
-            <tr class="{row_class}">
-                <td>{finding.status}</td>
-                <td>{finding.check_metadata.Severity}</td>
-                <td>{finding.check_metadata.ServiceName}</td>
-                <td>{finding.region}</td>
-                <td>{finding.check_metadata.CheckTitle}</td>
-                <td>{finding.resource_id.replace("<", "&lt;").replace(">", "&gt;").replace("_", "<wbr>_")}</td>
-                <td>{finding.check_metadata.Description}</td>
-                <td>{finding.check_metadata.CheckID.replace("_", "<wbr>_")}</td>
-                <td>{finding.status_extended}</td>
-                <td><p class="show-read-more">{finding.check_metadata.Risk}</p></td>
-                <td><p class="show-read-more">{finding.check_metadata.Remediation.Recommendation.Text}</p></td>
-                <td><a class="read-more" href="{finding.check_metadata.Remediation.Recommendation.Url}"><i class="fas fa-external-link-alt"></i></a></td>
-            </tr>
-            """
-    )
-
 
 def extract_findings_statistics(findings: list) -> dict:
     stats = {}
@@ -1111,132 +466,3 @@ def extract_findings_statistics(findings: list) -> dict:
 
     return stats
 
-
-def fill_html_overview_statistics(stats, output_filename, output_directory):
-    try:
-        filename = f"{output_directory}/{output_filename}{html_file_suffix}"
-        #  Read file
-        with open(filename, "r") as file:
-            filedata = file.read()
-
-        # Replace statistics
-        # TOTAL_FINDINGS
-        filedata = filedata.replace("TOTAL_FINDINGS", str(stats.get("findings_count")))
-        # TOTAL_RESOURCES
-        filedata = filedata.replace(
-            "TOTAL_RESOURCES", str(stats.get("resources_count"))
-        )
-        # TOTAL_PASS
-        filedata = filedata.replace("TOTAL_PASS", str(stats.get("total_pass")))
-        # TOTAL_FAIL
-        filedata = filedata.replace("TOTAL_FAIL", str(stats.get("total_fail")))
-        # Write file
-        with open(filename, "w") as file:
-            file.write(filedata)
-
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}] -- {error}"
-        )
-        sys.exit()
-
-
-def add_html_footer(output_filename, output_directory):
-    try:
-        filename = f"{output_directory}/{output_filename}{html_file_suffix}"
-        file_descriptor = open_file(
-            filename,
-            "a",
-        )
-        file_descriptor.write(
-            """
-               </tbody>
-            </table>
-        </div>
-    </div>
-    </div>
-    </div>
-    <!-- Table search and paginator -->
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js"
-        integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js"
-        integrity="sha384-1CmrxMRARb6aLqgBO7yyAxTOQE2AKb9GfXnEo760AUcUmFx3ibVJJAzGytlQcNXd"
-        crossorigin="anonymous"></script>
-    <!-- https://datatables.net/download/index with jQuery, DataTables, Buttons, SearchPanes, and Select //-->
-    <script type="text/javascript"
-        src="https://cdn.datatables.net/v/dt/jqc-1.12.4/dt-1.10.25/b-1.7.1/sp-1.4.0/sl-1.3.3/datatables.min.js"></script>
-    <script>
-        $(document).ready(function () {
-            // Initialise the table with 50 rows, and some search/filtering panes
-            $('#findingsTable').DataTable({
-                responsive: true,
-                // Show 25, 50, 100 and All records
-                lengthChange: true,
-                lengthMenu: [[25, 50, 100, -1], [25, 50, 100, "All"]],
-                searchPanes: {
-                    cascadePanes: true,
-                    viewTotal: true,
-                },
-                dom: 'Blfrtip',
-                language: {
-                    // To enable a filter button instead of the filter row
-                    searchPanes: {
-                        clearMessage: 'Clear Filters',
-                        collapse: { 0: 'Filters', _: 'Filters (%d)' },
-                        initCollapsed: true
-
-                    }
-                },
-                buttons: [
-                    {
-                        extend: 'searchPanes',
-                        config: {
-                            cascadePanes: true,
-                            viewTotal: true,
-                            orderable: false
-                        }
-                    }
-                ],
-                columnDefs: [
-                    {
-                        searchPanes: {
-                            show: true,
-                            pagingType: 'numbers',
-                            searching: true
-                        },
-                        // Show all filters
-                        targets: [0, 1, 2, 3, 5, 7]
-                    }
-                ]
-            });
-            var maxLength = 30;
-            // ReadMore ReadLess
-            $(".show-read-more").each(function () {
-                var myStr = $(this).text();
-                if ($.trim(myStr).length > maxLength) {
-                    var newStr = myStr.substring(0, maxLength);
-                    var removedStr = myStr.substring(maxLength, $.trim(myStr).length);
-                    $(this).empty().html(newStr);
-                    $(this).append(' <a href="javascript:void(0);" class="read-more">read more...</a>');
-                    $(this).append('<span class="more-text">' + removedStr + '</span>');
-                }
-            });
-            $(".read-more").click(function () {
-                $(this).siblings(".more-text").contents().unwrap();
-                $(this).remove();
-            });
-        });
-    </script>
-</body>
-
-</html>
-"""
-        )
-        file_descriptor.close()
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}] -- {error}"
-        )
-        sys.exit()
